IMPALA-823: Catalog Server does not handle Hive Metastore connection failures
When a HMS connection failed to open, an unchecked exception was being thrown. This
wasn't getting handled properly and was causing the loading threads to die. This fixes
the problem by ensuring the loading threads catch all types of exceptions and also
fixes the TableLoader to return an IncompleteTable should a HMS connection failure
occur.

Change-Id: I3b696fd8ef12aa6749b602324dcdfe4d27c935ee
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1609
Reviewed-by: Lenni Kuff <lskuff@cloudera.com>
Tested-by: jenkins