Fix Hdfs sink using wrong partition descriptor when partition keys are NULL
This bug was unconvered after the fix to IMPALA-624. If two partitions
have NULL as their first partition key, and an INSERT's partition clause
mentions that key statically as NULL, then the last partition processed
would be used, even if it was the partition.

The reason that this didn't come to light before was that the only way
that partitions differ from one another is either in their output
format, or in their location. We have no test for this situation with
different output formats, and prior to IMPALA-624 the location of all
partitions is dynamically determined from their partition keys. After
IMPALA-624, partitions with NULL keys have explicitly set location
fields, and since a random partition was being chosen, that location was
often incorrect.

This is covered by tests in alter_table.test.

Change-Id: I6035f092ccb9c8dc8c947af351642b51474747b6
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1666
Reviewed-by: Henry Robinson <henry@cloudera.com>
Tested-by: jenkins