Changes necessary for expr refactoring
This is some fairly mechanical refactoring that will be necessary for
the expr refactoring branch. It introduces the following changes:

- lookupSymbol() is moved from CreateFunctionStmtBase to Function
- HdfsFsFache and HdfsLibCache are removed from ExecEnv and made into
  standalone singletons
- Various 'is_fe_tests' variables are consolidated into one variable
  that is initialized early

This is paving the way for having the FE call into the BE on startup
to resolve builtins' symbols. The general idea is to simplify the
order various services are created, by having classes with fewer
dependencies initialized first.

Change-Id: I5d0e1325792a53680a9738e2e3e67fed6201299c
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1612
Tested-by: jenkins
Reviewed-by: Skye Wanderman-Milne <skye@cloudera.com>