Remove HdfsLibCache dtor
As far as I can tell, when the HdfsLibCache lived in the ExecEnv, its
destructor never got called, because there's no way for impalad to
exit its main function besides crashing or being killed. However, it
appears that static objects are properly destructed when the process
is killed, so the "Changes necessary for expr refactoring" commit
caused ~HdfsLibCache to be called, which results in a segfault related
to the JNI env being torn down as well.

Since we were leaking all HDFS connections before anyway, this patch
restores that behavior to avoid segfaults on shutdown.

Change-Id: Iecb0d6db39a4e5107a230b6957d0e77ddbe50c1f
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1697
Tested-by: jenkins
Reviewed-by: Skye Wanderman-Milne <skye@cloudera.com>