IMP-1306: Fix build scripts to properly generate Impala version info for packaging builds
The problem was that were were deleting the version.info file because the default
of gen_build_version.py recently changed from --noclean to --clean.

Also fixed a bug in the shell version generation and made debugging a bit easier
by dumping the contents of version.info whenever it is generated.

Change-Id: I764d01c9e46eed1bd39de79bf076c15afa599486
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1901
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: Lenni Kuff <lskuff@cloudera.com>