IMPALA-788: Loading an HBase table when HBase/Zookeeper are unavailable takes a long time to fail
This is because the default HBase/Zookeeper client timeouts are meant to be very long:

* If both Zookeeper and HBase are down, it will wait:
  hbase.client.retries.number(default 10) x hbase.rpc.timeout (default 60s)
* If ZK is up but HBase is down, but it has run at least once, then it waits:
  hbase.client.retries.number * some exponential backoff time (default sleep time is 1 second,
  backoff tables looks like this: 1, 1, 1, 2, 2, 4, 4, 8, 16, 32, 64).

In my experiments, it takes ~20-25minutes if HBase before the table loading will fail. If there
are many HBase tables this can block all loading threads.

The fix is to change the default timeout values to fail faster. These values were suggested
by somoene from the HBase team. With these values we will fail in ~1 minute. I am working with the CM team to get the defaults changed there as well.

Change-Id: I625e35af57374c72d50d03372d177624ce67694a
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1903
Reviewed-by: Nong Li <nong@cloudera.com>
Reviewed-by: Lenni Kuff <lskuff@cloudera.com>
Tested-by: Lenni Kuff <lskuff@cloudera.com>