Admission controller: Fix default parameter to QueuePlacementPolicy.assignAppToQueue()
We should be passing YarnConfiguration.DEFAULT_QUEUE_NAME as the default value to
QueuePlacementPolicy.assignAppToQueue(). This caused a problem when the placement
policy was Specified(create=true) where the returned name was "root." because it
literally created a pool with the name "" below root.

Change-Id: I41cc2848d78bf8bd308292e80d6f6c7d06437ae9
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1942
Reviewed-by: Nong Li <nong@cloudera.com>
Tested-by: jenkins
(cherry picked from commit 451627ffca6fffb562adcb18290bc7388d734e7f)
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1949
Reviewed-by: Matthew Jacobs <mj@cloudera.com>