IMPALA-900: Allow queries without a resource reservation to run with RM enabled
Coordinator fragments don't have any resources reserved for
them. Coordinator-only queries (like constant selects) therefore don't
have any resources reserved for them at all. This would trigger a DCHECK
that was a bit overzealous in PlanFragmentExecutor::Open().

Now rather than FLAGS_enable_rm we check whether the query resource
manager is NULL or not to decide whether to perform RM-related tasks for
queries. The query resource manager is set if and only if RM is enabled
and a reservation is present.

Change-Id: Ib1702bc4475d7cf633428f18596771a6e7f3eaa9
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1983
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: jenkins