Admission controller: Change default values for the "default pool"
The admission controller is configurable via Yarn fair scheduler allocation
and Llama configurations, but a "default pool" is used when these files are
not provided. When a pool is defined in a fair-scheduler.xml but no limits
are specified, the following Yarn/Llama default values are used: the max
number of concurrent requests is 20, the max queue size is 50, and the mem
limit is unlimited.

This changes the default values of the "default pool" limits so that the
limits are consistent with the defaults from Yarn/Llama.

Change-Id: Ic76ff550c18cc49353c72926591af46dcbe26ac7
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2006
Reviewed-by: Matthew Jacobs <mj@cloudera.com>
Tested-by: jenkins
(cherry picked from commit 1619d83e452e5b868d12e3934e9704fc5f16cac7)
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2118