IMPALA-908: Requests should fail gracefully if "user" is not set
We should more gracefully handle cases where a username is not
specified. Currently a Precondition will fail, but we should (a)
fall back to a default user and (b) have a way to enforce that
the username is specified and then return a more helpful error
when this occurs.

I've tested this with the jdbc driver with -require_user enabled
and disabled. When -require_user=true, we will return an error:

Using JDBC Driver Name: org.apache.hive.jdbc.HiveDriver
Connecting to: jdbc:hive2://localhost:21050/;auth=noSasl
Executing: select 1
java.sql.SQLException: User must be specified because -require_user is enabled.

Change-Id: I3312d13c0b9b269a4b0e789990995689137e9409
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2021
Reviewed-by: Lenni Kuff <lskuff@cloudera.com>
Tested-by: jenkins